pyproject.toml
[project]
name = "xdr-ioc-uploader"
version = "0.1.0"
description = "CLI to validate and upload IOCs (CSV/JSON) to Cortex XDR"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Rafael Guevara Hernádez", email = "rafagh00@gmail.com" }
]
dependencies = [
  "requests>=2.32.0",
  "pydantic>=1.10,<3",
  "python-dotenv>=1.0.1",
  "typer[all]>=0.12.3",
  "tenacity>=8.2.3",
  "rich>=13.7.1",
  "chardet>=5.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.5"
]
pandas = [
  "pandas>=2.2.2"
]

[project.scripts]
xdr-ioc-uploader = "xdr_ioc_uploader.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "B", "UP"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

